requirements.txt
numpy==1.20.3
opencv-python==4.5.2.54
pika==1.2.0
python-dotenv==0.17.1
